+++ src/pytest_mongo/__init__.py
@@ -18,4 +18,4 @@
 # along with pytest-mongo.  If not, see <http://www.gnu.org/licenses/>.
 """pytest-mongo's main module."""
 
-__version__ = '2.0.0'
+__version__ = '2.1.0'